+++ src/ext/index.js
@@ -1,0 +1,35 @@
+import getPropertyPanelDefinition from './property-panel';
+import getData from './data';
+
+const getExploration = (env) =>
+  env.flags.isEnabled('PS_18291_TABLE_EXPLORATION') && {
+    exploration: {
+      component: 'items',
+      items: {
+        columnHandler: {
+          component: 'column-handler',
+          search: true,
+          selectAll: true,
+        },
+      },
+      classification: {
+        tags: ['exploration'],
+        exclusive: true,
+      },
+    },
+  };
+
+export default function ext(env) {
+  return {
+    definition: getPropertyPanelDefinition(env),
+    ...getExploration(env),
+    data: getData(env),
+    support: {
+      export: true,
+      exportData: true,
+      snapshot: true,
+      viewData: false,
+      exploration: true,
+    },
+  };
+}