+++ src/table/types.ts
@@ -21,20 +21,23 @@ interface Action<T = any> {
   type: T;
 }
 
-export interface ActionPayload {
+export interface SelectPayload {
   cell: Cell;
   announce: Announce;
   evt: React.KeyboardEvent | React.MouseEvent;
 }
 
 export interface SelectAction extends Action<SelectionActions.SELECT> {
-  payload: ActionPayload;
+  payload: SelectPayload;
 }
-export interface SelectMultiStartAction extends Action<SelectionActions.SELECT_MULTI_START> {
-  payload: { cell: Cell };
+export interface SelectMouseDownAction extends Action<SelectionActions.SELECT_MOUSE_DOWN> {
+  payload: { cell: Cell; mouseupOutsideCallback(): void };
+}
+export interface SelectMouseUpAction extends Action<SelectionActions.SELECT_MOUSE_UP> {
+  payload: SelectPayload;
 }
 export interface SelectMultiAddAction extends Action<SelectionActions.SELECT_MULTI_ADD> {
-  payload: ActionPayload;
+  payload: SelectPayload;
 }
 export interface SelectMultiEndAction extends Action<SelectionActions.SELECT_MULTI_END> {}
 export interface ResetAction extends Action<SelectionActions.RESET> {}
@@ -45,7 +48,8 @@ export interface UpdateAllRowsAction extends Action<SelectionActions.UPDATE_ALL_
 
 export type SelectionActionTypes =
   | SelectAction
-  | SelectMultiStartAction
+  | SelectMouseDownAction
+  | SelectMouseUpAction
   | SelectMultiAddAction
   | SelectMultiEndAction
   | ResetAction
@@ -61,6 +65,7 @@ export interface SelectionState {
   api: ExtendedSelectionAPI;
   isSelectMultiValues: boolean;
   firstCell?: Cell;
+  mouseupOutsideCallback?(): void;
 }
 
 export interface ContextValue {
@@ -114,6 +119,7 @@ export interface HandleHeadKeyDownProps {
   layout: TableLayout;
   isInteractionEnabled: boolean;
   setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>;
+  areBasicFeaturesEnabled: boolean;
 }
 
 export interface HandleBodyKeyDownProps {
@@ -211,6 +217,7 @@ export interface TableHeadWrapperProps extends CommonTableProps {
   changeSortOrder: ChangeSortOrder;
   constraints: stardust.Constraints;
   translator: ExtendedTranslator;
+  areBasicFeaturesEnabled: boolean;
 }
 
 export interface TableBodyWrapperProps extends CommonTableProps {
@@ -228,6 +235,7 @@ export interface TableTotalsProps extends CommonTableProps {
   rootElement: HTMLElement;
   layout: TableLayout;
   selectionsAPI: ExtendedSelectionAPI;
+  areBasicFeaturesEnabled: boolean;
 }
 
 export interface PaginationContentProps extends CommonTableProps {