+++ src/table/utils/__tests__/handle-key-press.spec.ts
@@ -12,6 +12,7 @@ import {
 } from '../handle-key-press';
 import * as handleAccessibility from '../accessibility-utils';
 import * as handleScroll from '../handle-scroll';
+import * as handleCopy from '../copy-utils';
 import { Announce, Column, ExtendedSelectionAPI, Cell, TableLayout, TotalsPosition } from '../../../types';
 import { SelectionDispatch } from '../../types';
 import { KeyCodes } from '../../constants';
@@ -21,6 +22,11 @@ describe('handle-key-press', () => {
 
   beforeEach(() => {
     areBasicFeaturesEnabled = true;
+    jest.spyOn(handleCopy, 'default');
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   describe('shouldBubble', () => {
@@ -41,6 +47,7 @@ describe('handle-key-press', () => {
       isSelectionMode = false;
       keyboardEnabled = false;
       paginationNeeded = true;
+      areBasicFeaturesEnabled = true;
     });
 
     it('should return true when esc is pressed and isSelectionMode is false', () => {
@@ -274,6 +281,7 @@ describe('handle-key-press', () => {
         layout,
         isInteractionEnabled,
         setFocusedCellCoord,
+        areBasicFeaturesEnabled,
       });
 
     beforeEach(() => {
@@ -287,6 +295,7 @@ describe('handle-key-press', () => {
         target: {
           blur: jest.fn(),
           setAttribute: jest.fn(),
+          children: [],
         } as unknown as HTMLElement,
       } as unknown as React.KeyboardEvent;
       rootElement = {
@@ -345,6 +354,28 @@ describe('handle-key-press', () => {
       expect(changeSortOrder).not.toHaveBeenCalled();
       expect(setFocusedCellCoord).not.toHaveBeenCalled();
     });
+
+    it('should call copyCellValue on Head cell when the pressed keys are Ctrl and C keys', () => {
+      evt.key = KeyCodes.C;
+      evt.ctrlKey = true;
+      callHandleHeadKeyDown();
+      expect(handleCopy.default).toHaveBeenCalled();
+    });
+
+    it('should call copyCellValue on Head cell when the pressed keys are Meta and C keys', () => {
+      evt.key = KeyCodes.C;
+      evt.metaKey = true;
+      callHandleHeadKeyDown();
+      expect(handleCopy.default).toHaveBeenCalled();
+    });
+
+    it('should not call copyCellValue on Head cell when the flag is disabled', () => {
+      evt.key = KeyCodes.C;
+      evt.metaKey = true;
+      areBasicFeaturesEnabled = false;
+      callHandleHeadKeyDown();
+      expect(handleCopy.default).not.toHaveBeenCalled();
+    });
   });
 
   describe('handleTotalKeyDown', () => {
@@ -362,6 +393,7 @@ describe('handle-key-press', () => {
         target: {
           blur: jest.fn(),
           setAttribute: jest.fn(),
+          children: [],
         } as unknown as HTMLElement,
       } as unknown as React.KeyboardEvent;
       cellCoord = [1, 1];
@@ -375,7 +407,7 @@ describe('handle-key-press', () => {
     });
 
     it('should move the focus from the current cell to the next when arrow key down is pressed on a total cell', () => {
-      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode);
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
       expect(evt.preventDefault).toHaveBeenCalledTimes(1);
       expect(evt.stopPropagation).toHaveBeenCalledTimes(1);
       expect((evt.target as HTMLElement).setAttribute).toHaveBeenCalledTimes(1);
@@ -384,7 +416,7 @@ describe('handle-key-press', () => {
 
     it('should only call preventDefaultBehavior when isSelectionMode is true', () => {
       isSelectionMode = true;
-      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode);
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
       expect(evt.preventDefault).toHaveBeenCalledTimes(1);
       expect(evt.stopPropagation).toHaveBeenCalledTimes(1);
       expect(setFocusedCellCoord).not.toHaveBeenCalled();
@@ -392,10 +424,32 @@ describe('handle-key-press', () => {
 
     it('should take the default case when the pressed key is not an arrow key', () => {
       evt.key = KeyCodes.ENTER;
-      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode);
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
       expect(evt.preventDefault).toHaveBeenCalledTimes(1);
       expect(evt.stopPropagation).toHaveBeenCalledTimes(1);
       expect(setFocusedCellCoord).not.toHaveBeenCalled();
+    });
+
+    it('should call copyCellValue on Total cell when the pressed keys are Ctrl and C keys', () => {
+      evt.key = KeyCodes.C;
+      evt.ctrlKey = true;
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
+      expect(handleCopy.default).toHaveBeenCalled();
+    });
+
+    it('should call copyCellValue on Total cell when the pressed keys are Meta and C keys', async () => {
+      evt.key = KeyCodes.C;
+      evt.metaKey = true;
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
+      expect(handleCopy.default).toHaveBeenCalled();
+    });
+
+    it('should not call copyCellValue on Total cell when the flag is disabled', () => {
+      evt.key = KeyCodes.C;
+      evt.metaKey = true;
+      areBasicFeaturesEnabled = false;
+      handleTotalKeyDown(evt, rootElement, cellCoord, setFocusedCellCoord, isSelectionMode, areBasicFeaturesEnabled);
+      expect(handleCopy.default).not.toHaveBeenCalled();
     });
   });
 
@@ -443,6 +497,7 @@ describe('handle-key-press', () => {
           classList: {
             contains: () => isExcluded,
           },
+          children: [],
         } as unknown as HTMLElement,
       } as unknown as React.KeyboardEvent;
       rootElement = {
@@ -466,12 +521,7 @@ describe('handle-key-press', () => {
       areBasicFeaturesEnabled = true;
       jest.spyOn(handleAccessibility, 'focusSelectionToolbar').mockImplementation(() => jest.fn());
       jest.spyOn(handleAccessibility, 'announceSelectionState').mockImplementation(() => jest.fn());
-      jest.spyOn(handleAccessibility, 'copyCellValue');
       jest.spyOn(handleScroll, 'handleNavigateTop').mockImplementation(() => jest.fn());
-    });
-
-    afterEach(() => {
-      jest.clearAllMocks();
     });
 
     it('when press arrow down key on body cell, should prevent default behavior, remove current focus and set focus and attribute to the next cell', () => {
@@ -655,14 +705,14 @@ describe('handle-key-press', () => {
       evt.key = KeyCodes.C;
       evt.ctrlKey = true;
       runHandleBodyKeyDown();
-      expect(handleAccessibility.copyCellValue).toHaveBeenCalled();
+      expect(handleCopy.default).toHaveBeenCalled();
     });
 
     it('should call copyCellValue when the pressed keys are Meta and C keys', () => {
       evt.key = KeyCodes.C;
       evt.metaKey = true;
       runHandleBodyKeyDown();
-      expect(handleAccessibility.copyCellValue).toHaveBeenCalled();
+      expect(handleCopy.default).toHaveBeenCalled();
     });
 
     it('should not call copyCellValue when the flag is disabled', () => {
@@ -670,7 +720,7 @@ describe('handle-key-press', () => {
       evt.metaKey = true;
       areBasicFeaturesEnabled = false;
       runHandleBodyKeyDown();
-      expect(handleAccessibility.copyCellValue).not.toHaveBeenCalled();
+      expect(handleCopy.default).not.toHaveBeenCalled();
     });
 
     it('when other keys are pressed, should only prevent default behavior', () => {
@@ -738,8 +788,6 @@ describe('handle-key-press', () => {
       isSelectionMode = true;
       jest.spyOn(handleAccessibility, 'focusSelectionToolbar').mockImplementation(() => jest.fn());
     });
-
-    afterEach(() => jest.clearAllMocks());
 
     it('should call focusSelectionToolbar when isSelectionMode is true and tab is pressed', () => {
       handleLastTab(evt, isSelectionMode, keyboard);